Extract the mutually exclusive component to its own file
